27 prevent users from removing the first question

+++ components/CreateQuiz/index.tsx
@@ -178,19 +178,35 @@ const CreateQuiz: React.FC<CreateQuizProps> = ({ id }) => {
     setQuestionList(data.data.questionList);
   };
 
-  const handleDeleteQuestion = async (item: Quiz) => {
-    const res = await fetch("/api/remove_quiz/" + item._id + "/" + id, {
-      method: "DELETE",
-    });
-    if (res.status === 401) {
-      await signOut();
-      router.push("/");
-      return;
+  const checkIsFirstQuestion = (id: string): boolean => {
+    // const question =  questionList.map((question, i )=>{if (question._id === id) {return question}});
+
+    let question: Quiz;
+
+    for (let i = 0; i < questionList.length; i++) {
+      if (questionList[i]._id === id) {
+        question = questionList[i];
+      }
     }
+    console.log(question);
+    return question.isFirst;
+  };
+  const handleDeleteQuestion = async (item: Quiz) => {
+    if (!checkIsFirstQuestion) {
+      const res = await fetch("/api/remove_quiz/" + item._id + "/" + id, {
+        method: "DELETE",
+      });
+      if (res.status === 401) {
+        await signOut();
+        router.push("/");
+        return;
+      }
 
-    const data = await res.json();
-
-    setQuestionList(data.data.questionList);
+      const data = await res.json();
+      setQuestionList(data.data.questionList);
+    } else {
+      console.log("Cannot delete the first question");
+    }
   };
   const handleAddAnswer = async () => {
     switch (panelStatus) {
@@ -216,6 +232,20 @@ const CreateQuiz: React.FC<CreateQuizProps> = ({ id }) => {
       default: {
         break;
       }
+    }
+  };
+
+  const handleDeleteTroubleshoot = async () => {
+    const res = await fetch("/api/remove_troubleshoot/" + id, {
+      method: "DELETE",
+    });
+    if (res.status === 401) {
+      await signOut();
+      router.push("/");
+      return;
+    }
+    if (res.status === 200) {
+      router.push("/create_troubleshoot");
     }
   };
 
@@ -252,6 +282,12 @@ const CreateQuiz: React.FC<CreateQuizProps> = ({ id }) => {
               >
                 <div className="border-2 rounded p-2">
                   <h1>{troubleShootTitle}</h1>
+                  <Button
+                    title="Delete Troubleshoot"
+                    onClick={() => {
+                      handleDeleteTroubleshoot();
+                    }}
+                  />
                 </div>
                 {/* New Question */}
                 <div className="flex flex-row ">
@@ -301,13 +337,19 @@ const CreateQuiz: React.FC<CreateQuizProps> = ({ id }) => {
                                 setVisble(true);
                               }}
                             />
-                            <Button
-                              title={"Delete"}
-                              onClick={() => {
-                                // setSelectedQuestion(item);
-                                handleDeleteQuestion(item);
-                              }}
-                            />
+                            {!item.isFirst ? (
+                              <Button
+                                title={"Delete"}
+                                onClick={() => {
+                                  // setSelectedQuestion(item);
+                                  handleDeleteQuestion(item);
+                                }}
+                              />
+                            ) : (
+                              <div>
+                                This is the first Question, cannot be deleted
+                              </div>
+                            )}
                           </div>
                         </div>
                       );
